Moving privacy exception to common exceptions to avoid circular dependency
