+++ src/fides/api/common_exceptions.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from typing import List
+from typing import Dict, List, Optional
 
 from fastapi import HTTPException, status
 from starlette.status import (
@@ -371,6 +371,15 @@ class InvalidScopeError(HTTPException):
                 "valid_scopes": SCOPES,
             },
         )
+
+
+class PrivacyRequestError(Exception):
+    """Base exception for privacy request operations."""
+
+    def __init__(self, message: str, data: Optional[Dict] = None):
+        self.message = message
+        self.data = data
+        super().__init__(message)
 
 
 class KeyOrNameAlreadyExists(Exception):
